requirements.txt
google-generativeai==0.3.2
supabase==2.0.2
